Update to Build 023 (Alpha 0.3)
Added some new script features:
ForkQuestion : This allows for the script to ask the player a question
and then fork off into different scripts according to what the player
answered.
((Its kinda buggy right now, so don't expect it to work))

Added a couple lines and a script to Emi's scripts.