Update to Build 027 (Alpha 0.31) *Katawa Shoujo!*
Just a random update.  Added nothing to engine or to normal gameplay.
But, there is a random snippet of Katawa Shoujo script in the code
hidden only available via debugging.  I did this to see if Katawa could
run natively in my engine.  It can with a bit of tweaking, some of our
commands are different and mine requires a comma at the end of every
line.

It was fun doing it.