Update to Build 029 (Alpha 0.35)
This update allows the authors to test run their scripts instantly on
game run.  This is great for the authors that are not comfortable in
adding Events in the main game loop.  It's also good because it allows
for even more experienced devs to run their script without having to
trigger it in game.

In order to run your script on startup you must do 2 things.
1) call the script "debug".
2) Find and click the hidden debug button on the start menu.

The debug button is the TurtleSim logo.  Just click it.